Crate walox

The lox language.

Modules

analysis
This module contains static analysis tools for lox AST that can detect specific errors rather than encountering them at runtime.

ast
Abstract syntax tree definitions and utilities

compiler
Utilities for the single-pass compiler

interpreter
Tree-walking interpreter for the AST

parser
Parser implementation

scanner
Tools for working with lox source code

span
Common structures for reporting and selecting information from source code

vm
Virtual machine implementation