Enum walox::ast::ExprKind

pub enum ExprKind {
    Binary(BinaryExpr),
    Grouping(GroupingExpr),
    Literal(LiteralExpr),
    Unary(UnaryExpr),
    Logical(LogicalExpr),
    Var(VarExpr),
    Assign(AssignExpr),
    Call(CallExpr),
    Get(GetExpr),
    Set(SetExpr),
    This(ThisExpr),
    Super(SuperExpr),
}

Different types of expressions

Variants

Binary(BinaryExpr)

A binary operation

Grouping(GroupingExpr)

Different types of binary operations

Literal(LiteralExpr)

A literal value

Unary(UnaryExpr)

A unary operation

Logical(LogicalExpr)

A logical operation

Var(VarExpr)

A variable reference

Assign(AssignExpr)

An assignment to an existing variable

Call(CallExpr)

An expression that calls a function with the supplied arguments

Get(GetExpr)

An expression that access a named property on an object

Set(SetExpr)

An expression that sets the value of a named property on an object

This(ThisExpr)

An expression that references the value of the instance that was the original owner of the method containing this.

Super(SuperExpr)

An expression that acts as a property access on the superclass