Module walox::interpreter

Tree-walking interpreter for the AST

Modules

native_funcs
Commonly used functions that the host environment provides to lox programs

Structs

Class
A class is an extensible collection of methods that is used to create objects

Environment
The set of bindings that are present in lexical scopes during execution.

Instance
An instantiation of a class

Interpreter
The AST interpreter

InvalidOperationForType
An error that occurs when performing an operation between incompatible types

NativeFunction
A function that is defined by the interpreter

UserFunction
A user defined function

Enums

FunctionType
Different types of functions

RuntimeControlFlow
Errors that can occur during the course of interpretation

RuntimeException
Errors that can occur during the course of interpretation

Value
A lox value