Module walox::parser

Parser implementation

Enums

ParseError
An error that can occur during the parsing process

Functions

addition
Parse an addition or subtraction expression

and
Parse an AND logical expression

assignment
Parse an assignment expression

block
Parse a block of statements

call
Parse a function call expression

class_declaration
Parse a class declaration

comparison
Parse a comparison expression

declaration
Parse a declaration or fall through to a normal statement

equality
Parse an equality expression

expr_statement
Parse an expression statement

expression
Parse an expression

for_statement
Parse a for statement

function_declaration
Parse a function declaraction

if_statement
Parse an if statement

multiplication
Parse a multiplication or division expression

or
Parse an OR logical expression

parse
Parse lox source

primary
Parse a primary expression: a literal or a grouping expression

print_statement
Parse a print statement

return_statement
Parse a return statement

statement
Parse a statement

synchronize
Take the current state of the Cursor and attempt to fast-forward until a reasonable parse boundary is found

unary
Parse a unary operation expression

var_declaration
Parse a variable declaraction

while_statement
Parse a while statement