Module walox::scanner

Tools for working with lox source code

Structs

Cursor

A struct which manages the state of the Token iterator and provides common utilities

MissingTokenError

A Cursor expected a TokenType that was not found

Scanner

The Scanner takes raw text input and produces a sequence of Tokens

Token

A Token is an instance of a chunk of text with meaning, which cannot be subdivided

TokenTryFromError

An error that occurs when trying to convert from u8 to TokenType.

Enums

Literal

A literal value embedded in the source code

ScanError

Errors that can occur during the scanning process

TokenType

The TokenType represent the type of the chunks of text