Module walox::vm

Virtual machine implementation

Structs

Chunk
An immutable structure that contains compiled Instructions and other relevant data needed to execute a program.

ChunkBuilder
A builder structure that represents a Chunk in the process of being built.

ChunkIter
An iterator over all the instructions in a Chunk.

Heap
A memory region that contains Values, separate from the stack.

Instruction
An Instruction is the basic unit of execution in the lox virtual machine.

Object
An opaque object reference.

ObjectBase
The base fields of all Objects

StringObject
An immutable String object

TryFromByteError
The error type returned when conversion from a byte to an opcode fails.

VM
The virtual machine that executions Instructions

Enums

ChunkError
Errors that occur while iterating overing Instructions.

ObjectType
The type of an Object

OpCode
Virtual machine instruction type

RuntimeError
Errors that can occur during the execution of bytecode.

Value
The value type of the virtual machine

Traits

ConcreteObject
This trait encompasses the behavior of all concrete object types, i.e. all Objects that do not require more casting to be functional.