Enum walox::vm::OpCode

#[repr(u8)]
pub enum OpCode {

    Constant,
    Nil,
    True,
    False,
    Equal,
    Greater,
    Less,
    Add,
    Subtract,
    Multiply,
    Divide,
    Not,
    Negate,
    Print,
    Pop,
    Return,
    GetGlobal,
    DefineGlobal,
    SetGlobal,
    GetLocal,
    SetLocal,
    JumpIfFalse,
    Jump,
    Loop,

}

Virtual machine instruction type

Variants

Constant

Load a constant onto the stack

Nil

Push a constant Value::Nil onto the stack.

True

Push a constant Value::Bool(true) onto the stack.

False

Push a constant Value::Bool(false) onto the stack.

Equal

Compare two Values and return true if they are equal.

Greater

Compare two Value::Number and return true if the lhs is greater.

Less

Compare two Value::Number and return true if the lhs is less.

Add

Add two Values

Subtract

Subtract two Values

Multiply

Multiple two Values

Divide

Divide two Values

Not

Boolean inverse

Negate

Numeric inverse

Print

Pop the top value off the stack and write it to stdout

Pop

Pop the top value off the stack and do nothing with it

Return

Return from the current function

GetGlobal

Read a global variable’s value

DefineGlobal

Define a global variable with an initial value

SetGlobal

Write a new value to an existing global variable

GetLocal

Read a local variable

SetLocal

Set a local variable

JumpIfFalse

Increment the ip variable if the top of the stack is falsey

Jump

Unconditionally increment the ip variable

Loop

Unconditionally decrement the ip variable

Implementations

impl OpCode

pub const JUMP_OP_ARGUMENT_SIZE: usize = 2usize

The number of bytes for the jump operation arguments.

pub fn arguments_size(&self) -> usize

Returns the number of bytes of extra information needed to execute the instruction.

pub fn try_from_byte(byte: u8) -> Result<Self, TryFromByteError>

Attempt to onvert from a byte representation to the OpCode enum.

Errors

This function will return an error if the given byte is not a valid OpCode.

pub unsafe fn try_from_byte_unchecked(byte: u8) -> Self

Convert from a byte representation to the OpCode enum.

Safety

This byte must represent a valid OpCode.