Trait walox::vm::ConcreteObject

pub trait ConcreteObject: Display + PartialEq {
    const TYPE: ObjectType;

    fn upcast(own: *mut Self) -> OpaqueObject { ... }
}

This trait encompasses the behavior of all concrete object types, i.e. all Objects that do not require more casting to be functional.

Required Associated Constants

const TYPE: ObjectType

The ObjectType of the ConcreteObject.

Provided Methods

fn upcast(own: *mut Self) -> OpaqueObject

Take a mutable pointer to a ConcreteObject value, and produce an Object that references that initial value