Trait walox::ast::visit::Visitable

pub trait Visitable: Debug {
    fn super_visit_with<V: Visitor>(&self, visitor: &mut V) -> V::Output;

    fn visit_with<V: Visitor>(&self, visitor: &mut V) -> V::Output { ... }
}

A type which is able to be traversed by a Visitor

Required Methods

fn super_visit_with<V: Visitor>(&self, visitor: &mut V) -> V::Output

Perform the default visiting of this piece of data

Provided Methods

fn visit_with<V: Visitor>(&self, visitor: &mut V) -> V::Output

Visit the this piece of data with the given Visitor