Enum walox::scanner::TokenType

#[repr(u8)]
pub enum TokenType {

    LeftParen,
    RightParen,
    LeftBrace,
    RightBrace,
    Comma,
    Dot,
    Minus,
    Plus,
    Semicolon,
    Slash,
    Star,
    Bang,
    BangEqual,
    Equal,
    EqualEqual,
    Greater,
    GreaterEqual,
    Less,
    LessEqual,
    Identifier,
    String,
    Number,
    And,
    Class,
    Else,
    False,
    Fun,
    For,
    If,
    Nil,
    Or,
    Print,
    Return,
    Super,
    This,
    True,
    Var,
    While,
    Error,

}

The TokenType represent the type of the chunks of text

Variants

LeftParen

(

RightParen

)

LeftBrace

{

RightBrace

}

Comma

,

Dot

.

Minus

-

Plus

+

Semicolon

;

Slash

/

Star

*

Bang

!

BangEqual

!=

Equal

=

EqualEqual

==

Greater

>

GreaterEqual

>=

Less

<

LessEqual

<=

Identifier

An identifier is a name, such as abc, a23, brownCowNow

String

A piece of embedded text in the source, "Stinsdlksjdlfkjsd"

Number

A numerical value, 123, 1.234, 0.0

And

and

Class

class

Else

else

False

false

Fun

fun

For

for

If

if

Nil

nil

Or

or

Print

print

Return

return

Super

super

This

this

True

true

Var

var

While

while

Error

Error token, used to transmit error through Token iterator

Implementations

impl TokenType

pub fn keywords() -> HashMap<&'static str, TokenType>

Return the mapping from keyword strings to TokenTypes.

pub fn length(self) -> Option<usize>

Return the length of the token if it is known statically