Module walox::vm

Virtual machine implementation

Structs

CallFrame

The call frame tracks the live information needed for executing a single function call

Chunk

An immutable structure that contains compiled Instructions and other relevant data needed to execute a program.

ChunkBuilder

A builder structure that represents a Chunk in the process of being built.

ChunkIter

An iterator over all the instructions in a Chunk.

FunctionObject

A function defined in the lox source code

Heap

A memory region that contains Values, separate from the stack.

Instruction

An Instruction is the basic unit of execution in the lox virtual machine.

ObjectBase

The base fields of all Objects

ObjectRef

A non-opaque object reference.

OpaqueObject

An opaque object reference.

StringObject

An immutable String object

TryFromByteError

The error type returned when conversion from a byte to an opcode fails.

VM

The virtual machine that executions Instructions

Enums

ChunkError

Errors that occur while iterating overing Instructions.

ObjectType

The type of an Object

OpCode

Virtual machine instruction type

RuntimeError

Errors that can occur during the execution of bytecode.

Value

The value type of the virtual machine

Traits

ConcreteObject

This trait encompasses the behavior of all concrete object types, i.e. all Objects that do not require more casting to be functional.